Unify input handling to use same model
Previously, the depth/rudder joysticks are used to give a target rate of change value (i.e. depth speed or angular velocity), which is reached overtime with constant acceleration. The engines joystick, similarly, gave a target velocity reached overtime with constant acceleration. However, this is not realistic, so this has been updated to instead give a target acceleration, reached overtime with constant rate of change of acceleration, to unify the input model.